build: share `main_fd_inheritance_windows.go`
